Move Database Providers to DatabaseModule
